<finnal>

+++ shell-docker/rm-all-container.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 container=$(docker ps | awk '{print $1}' | sed '1d')
-docker stop $container
-gg=$(docker ps -a | awk '{print $1}' | sed '1d')
-docker rm $gg
+if [ -n "$container" ];then
+     docker stop $container
+     gg=$(docker ps -a | awk '{print $1}' | sed '1d')
+     docker rm $gg
+fi
+
+
+
+
+
